+++ src/llvm/load-constant-recorder.cc
@@ -6,13 +6,14 @@
 namespace v8 {
 namespace internal {
 namespace tf_llvm {
-int64_t LoadConstantRecorder::Register(int64_t magic,
-                                       LoadConstantRecorder::Type type,
-                                       int rmode) {
-  int result_magic = magic;
+uintptr_t LoadConstantRecorder::Register(uintptr_t magic,
+                                         LoadConstantRecorder::Type type,
+                                         int rmode) {
+  uintptr_t result_magic = magic;
   switch (type) {
     case kRelocatableInt32Constant:
-      result_magic = magic | (static_cast<int64_t>(type) << 16) | (rmode << 24);
+      result_magic =
+          magic | (static_cast<uintptr_t>(type) << 16) | (rmode << 24);
       break;
     default:
       break;
@@ -22,7 +23,7 @@ int64_t LoadConstantRecorder::Register(int64_t magic,
 }
 
 const LoadConstantRecorder::MagicInfo& LoadConstantRecorder::Query(
-    int64_t magic) const {
+    uintptr_t magic) const {
   auto found = map_.find(magic);
   EMASSERT(found != map_.end());
   return found->second;
